+++ tests/Assets/Providers/TestingServiceProvider.php
@@ -1,11 +1,16 @@
 <?php
 
-namespace Sfneal\Casts\Tests\Providers;
+namespace Sfneal\Casts\Tests\Assets\Providers;
 
 use Illuminate\Support\ServiceProvider;
 
 class TestingServiceProvider extends ServiceProvider
 {
+    /**
+     * Bootstrap any MockModel services.
+     *
+     * @return void
+     */
     public function boot()
     {
         // Publish migration file (if not already published)