+++ tests/TestCase.php
@@ -3,10 +3,11 @@
 namespace Sfneal\Casts\Tests;
 
 use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Foundation\Application;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Orchestra\Testbench\TestCase as OrchestraTestCase;
-use Sfneal\Casts\Tests\Models\People;
-use Sfneal\Casts\Tests\Providers\TestingServiceProvider;
+use Sfneal\Casts\Tests\Assets\Models\People;
+use Sfneal\Casts\Tests\Assets\Providers\TestingServiceProvider;
 
 class TestCase extends OrchestraTestCase
 {
@@ -17,14 +18,26 @@ class TestCase extends OrchestraTestCase
      */
     public $models;
 
+    /**
+     * Register package service providers.
+     *
+     * @param Application $app
+     * @return array|string
+     */
     protected function getPackageProviders($app)
     {
         return TestingServiceProvider::class;
     }
 
+    /**
+     * Define environment setup.
+     *
+     * @param Application $app
+     * @return void
+     */
     protected function getEnvironmentSetUp($app)
     {
-        include_once __DIR__.'/migrations/create_people_table.php.stub';
+        include_once __DIR__.'/Assets/migrations/create_people_table.php.stub';
 
         (new \CreatePeopleTable())->up();
     }